[NGR-1127] Fixing Pr comments

+++ app/uk/gov/hmrc/ngrraldfrontend/controllers/HowMuchIsTotalAnnualRentController.scala
@@ -24,7 +24,7 @@ import uk.gov.hmrc.ngrraldfrontend.config.AppConfig
 import uk.gov.hmrc.ngrraldfrontend.models.components.NGRRadio.buildRadios
 import uk.gov.hmrc.ngrraldfrontend.models.components.NavBarPageContents.createDefaultNavBar
 import uk.gov.hmrc.ngrraldfrontend.models.forms.HowMuchIsTotalAnnualRentForm
-import uk.gov.hmrc.ngrraldfrontend.models.forms.HowMuchIsTotalAnnualRentForm.{form}
+import uk.gov.hmrc.ngrraldfrontend.models.forms.HowMuchIsTotalAnnualRentForm.form
 import uk.gov.hmrc.ngrraldfrontend.models.registration.CredId
 import uk.gov.hmrc.ngrraldfrontend.repo.RaldRepo
 import uk.gov.hmrc.ngrraldfrontend.utils.Constants.{renewedAgreement, surrenderAndRenewal}
@@ -67,7 +67,7 @@ class HowMuchIsTotalAnnualRentController @Inject()(howMuchIsTotalAnnualRentView:
         rentAmount =>
           raldRepo.insertAnnualRent(
             credId = CredId(request.credId.getOrElse("")),
-            rentAmount = rentAmount.annualRent.toInt
+            rentAmount = rentAmount.annualRent
           )
           Future.successful(Redirect(routes.WhatTypeOfLeaseRenewalController.show.url))
       )
